requirement.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-tinymce==3.5.0
psycopg2-binary==2.9.5
python-decouple==3.6
sqlparse==0.4.3
